requirements.txt
absl-py==2.1.0
astunparse==1.6.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
flatbuffers==25.1.24
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.12.1
idna==3.10
immutabledict==4.2.1
jax==0.4.33
jaxlib==0.4.33
joblib==1.4.2
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
ml_collections==1.0.0
namex==0.0.8
networkx==3.4.2
nltk==3.9.1
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
protobuf==5.29.3
Pygments==2.19.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
scipy==1.15.1
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
textblob==0.19.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
